Mais algumas correções nos DAO

+++ src/main/java/ufg/pw/projeto_restaurante/model/utils/endereco/dao/EnderecoDao.java
@@ -3,7 +3,7 @@ package ufg.pw.projeto_restaurante.model.utils.endereco.dao;
 import ufg.pw.projeto_restaurante.model.utils.endereco.Endereco;
 import ufg.pw.projeto_restaurante.model.utils.persistencia.GenericDao;
 
-public class EnderecoDao extends GenericDao<Endereco, Integer> {
+public class EnderecoDao extends GenericDao<Endereco, Long> {
 	public EnderecoDao() {
 		super(Endereco.class);
 	}
